+++ crates/database/migration/Cargo.toml
@@ -6,14 +6,14 @@ publish = false
 
 [dependencies]
 alloy-primitives = { workspace = true, features = ["serde"] }
-async-std = { version = "1", features = ["attributes", "tokio1"] }
-futures.workspace = true
-reqwest = { workspace = true, features = ["stream"] }
 eyre.workspace = true
-csv = "1.3.1"
+futures.workspace = true
+indicatif = "0.17.11"
+reqwest = { workspace = true, features = ["stream"] }
+reqwest-retry = "0.7.0"
+reqwest-middleware = "0.4.2"
 sea-orm = { workspace = true, features = ["sqlx-sqlite", "runtime-tokio-native-tls", "macros"] }
-serde = { workspace = true, features = ["derive"] }
-tokio = { workspace = true, features = ["full"] }
+sha2 = "0.10.9"
 tracing.workspace = true
 
 [dependencies.sea-orm-migration]
