--- crates/database/migration/src/main.rs
@@ -1,8 +1,0 @@
-use sea_orm_migration::prelude::*;
-
-#[async_std::main]
-async fn main() {
-    tracing::info!(target: "scroll::migration", "Running database migrations.");
-    cli::run_cli(scroll_migration::Migrator).await;
-    tracing::info!(target: "scroll::migration", "Database migrations complete.")
-}
